repo reset. also, infinite loop somewhere?

+++ assignment3/new_code/schedule.cpp
@@ -9,7 +9,7 @@
 #define TIME_MID 16
 #define TIME_FACTOR 10
 
-double Player::get_time_schedule() {
+float Player::get_time_schedule() {
 	//I should return a small number initially.
 	if ( plies< TIME_EARLY ) { //first 8 plies.
 		return ((remaining_time>1.0)?(1.0):(remaining_time/TIME_FACTOR)); //1 second
